backend/requirements.txt
django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
gunicorn==22.0.0
requests==2.32.3

